Refactor router system and add declaritive page construction
Routes are moved into modules to avoid a monolithic base router that's
overflowing with the various routes from all the different modules.
Instead of using a subrouter, use a generator function that will
return a set of key-value pairs containing the relevants components of
any given page, ie layout, crumbs, views, etc. This allows to
drastically reduce the amount of boilerplate code needed to populate
the various parts of the page, while still providing the flexibility
for new layouts and other features to be added as needed.

The target div information for where views are to be rendered has been
moved back into the modules themselves, as creating helper functions
around setView for every potential target view location (main content,
sidebar, tabs, api bar, menu bar, etc etc) quickly becomes more
complex and brittle than its worth. Furthermore, as long as the module
defines what layout it expects to be rendered into, it will have
knowledge of the relevant divs and can use those accordingly. It could
be useful to revisit abstracting out the layout and target div
locations from modules, but at this point its a clunkier solution.

Views should now extend from the FauxtonAPI.View object, which at this
point adds one function "establish" that allows a view to return an
array of promises that are necessary to be completed before it is
ready to do a final render. This means we remove substantial
boilerplate for fetching some data and then rendering the appropriate
view when it finally returns, and also allows us to abstract out the
bookkeeping of those nested views into the dashboard object so that we
can easily render different views in the layout when they're ready.

The Route generator also allows for an establish function that will
let you require any needed data to be shared between views in a page,
or could possibly be used for user authentication, or any other
control flow requiring a delayed action to determine how to proceed.

The dashboard helper methods setDashboardContent and related are
being replaced by setView and renderView, which basically dispatch to
the layout for setting and rendering views, however, the dashboard
will keep a reference to the nested view returned by layout.setView
allowing you to easily reference that view for rendering at a later
point, again, removing substantial boilerplate assoicated with holding
references to nested views until the associated data has been fetched.