WIP: flesh out doc CRUD, add notifications, fix glyphs
Start working on making the transitions between all docs and docs
functional. Added in logic for save docs to check for valid code and
either add a fail warning or saving notification.

Added in notion of a pageUrl, which is a way for a model to define
where it is located in terms of a REST resource. There is a fair bit
of overlap with the url function, but not always, and its relative vs
absolute. I'm not a huge fan of the way pageUrl is setup and
constructed right now, but we do need a way for a modal to define
where its accessible in the app. Ie given a database or a doc, you can
provide a route compatible url that will get you to the relevant page
for that item.

We should also consider restructuring the application urls to be more
restful and not necessarily map directly to the couchdb urls, as we
have the api bar down on the bottom which does a much better job.
Things like listing all databases should potentially just be
/databases in the url as /databases/_all_dbs is akward, and becomes
tricky to distinguish between /databases/_design%2Fasdf.

Another related option is setting up a route
/databases/_design/:ddoc_id that will allow us to explicitly handle
design docs and we can use a different default view. For instance, on
normal docs, I prefer to see the code editor JSON, but on ddocs I
would like to see a higher level representation.

I temporarily added in div.view.show to make the styling and buttons
work on the _all_docs list, but we should replace that and move away
from that model.

Also fixed image copying in the build so the glyph icons work now, and
any other images will now be there as well.

Added a notification API allowing you to set a msg, notification type,
and optionally specify a target or by default just use a global notifier.