Fix inappropriate display of "Document saved successfully"
Background: If a doc is already loaded, the "Document
saved successfully" message appears when it is redisplayed
in the doc editor.

This patch modifies updateValues to check for changedAttributes
in the model and the notification is only displayed if there
are changed attributes.

Note that this correctly handles the case where a document is
saved that has no actual changes, as the _rev will be
different and so changedAttributes will still be truthy.