Fauxton: use local version of font awesome
We are currently referencing to Font-Awesome on a CDN in the
variables.less, which is basically nice, but some users of
CouchDB are firewalled at work and can just use the local
network.

Additionally offline people without internet can't use Fauxton
if the font is loaded via CDN.

This also removes the cache-buster for imgs (GET-Params with
version) in our routing for the local files of the dev-server
to make it work, in the future we might want to use a real
router module for the long if/else block.

Closes: COUCHDB-2238