Remove `rightHeader` from changes view for now
We have a separate filter function in this view. It probably
makes sense to integrate this filter into the right header
function later, but currently the rightHeader with the search for
documents and the query-options-tray does not work with changes and
duplicates the functionality of the filter in this context.

Closes COUCHDB-2397