Fix rendering on view-list after creating a view
AllDocsList is lazily adding subviews in the `beforeRender` method,
which causes that it will throw an error for all methods that
access these subviews from outside before explicitly rendering
the view.

As the beforeRender is adding paging, subviews, and the establish
takes care of our collection, we are fine with just rendering.

After we got this working, we can navigate without a
`trigger: false` and are getting a view on the indexed doc after
creating the view.