Fix for page body widths being incorrect on Stats
As noted in the ticket, the widths of the main body of the Stats and Accounts
page were incorrect. On the stats page, the rightmost column was offscreen; on the
Accounts page the text gets cut off.

This ticket fixes the stats page. Tthe Create Admins page still has the text get
cut off. This is unrelated to thisissue - I'll fix that in a separate ticket
(COUCHDB-2417).

Closes COUCHDB-2428