Fix for page body widths being incorrect on Stats #2
This is the second version of this ticket that doesn't contain whitespace fixes.

As noted in the ticket, the widths of the main body of the Stats and Accounts
page were incorrect. On the stats page, the rightmost column was offscreen; on the
Accounts page the text gets cut off.

This ticket fixes the stats page. The Create Admins page still has the text get
cut off. This is unrelated to this issue - I'll fix that in a separate ticket
(COUCHDB-2417).

Closes COUCHDB-2428