Unify one_pane and one_pane_db again
We recently introduced a new template `one_pane_db` which had
the rightHeader for the list of DBs. In retrospective it turns
out that also other views need the rightHeader markup to
populate it with elements.

Views that don't need the rightHeader element should not get
influenced so this removes some complexity and avoids that the
templates diverge more in the future.