Refactor Nightwatch
This ticket makes a few changes to the way Nightwatch is set up:
1. the main Nightwatch settings (username, password, login URL, key
nightwatch setup file locations) are moved to the settings.json
file to allow them to be defined a single global place and overridden
if need be.
2. it now dynamically determines the list of tests to run based on
content of settings.json.
3. hardcoded username, password & URL moved from helpers.js and
loginToGUI.js and now loaded from settings.json file

Closes COUCHDB-2481