Componentize footer
Take the elements that were all baked into all-docs-list
and inject them into the all-docs-list. This enables us to have
a separate footer component in the layout which can be initialized
without the all-docs-list viewobject.

This enables us to remove the `position: fixed` from the footer
and use `position: absolute`. Without having the element
`position: fixed;` we do not need to calculate it's width via JS
and can apply a css based pane layout.

Part of: COUCHDB-2471