New waitForAttribute nightwatch helper
This test allows us to not just wait for an element, but for an
element attribute containing a certain value. This can be helpful in a
number of scenarios, like when a JS event changes the content of a page
but the DOM structure remains largely the same (e.g. a form).