Redirect to login page on access denied
An error message is displayed in Fauxton telling the user that
access is denied. While it tells them to select the login link to
log in (which does work), this is not a good user experience.

In this case, the user should be presented with the login and not
an error message.

closes COUCHDB-2435