Full page doc editor fixes
It was noticed that with larger documents we needed the horizontal
scrollbar to appear, since some users don't have right-left
scrolling mice/trackpads and relying exclusively on the keyboard
wasn't sufficient. This adds the horizontal scrollbar only when
needed, and at the very bottom of the screen, regardless of the
height of the editor, to make it look ok.