header-control-bar: fix select-all-state
When I select all documents
And there are less documents than I have set to display per page
Then I want the select-all-button to be selected/disabled

As I checked for the setting of the page compared to the selected
documents it can happen that if we have not enough documents to
show on the page that the status of the select-all was wrong.

To test:
if you have database with 8 documents, and show 5 documents at a
time, view the second page (doc 6-8), and hit 'select all', notice
that the 'select all' and 'unselect all' are both grayed out.

Added test to avoid regressions in the future and to prove my fix