cleanup: remove useRAF leftovers
Garren and me tried to use the latest layoutmanager in summer
2014 - the experiment lead to performance improvements regarding
the rendering of our application, but we had to disable it as we
as it lead to timing issues.

As we are moving to React it does not make sense to me to invest
further time into the issues we had in summer.

This commit removes the leftovers of the experiment that are not
needed any more and cleans up.

See https://github.com/apache/couchdb-fauxton/pull/29