tooling: automatically test for code style
Currently taking a look if our js coding guidelines are followed
is done manually during the main code review of a PR. This takes
a lot of time for the reviewer and often results in 1-2 more
iterations until the patch is landed, just taking the style issues
into account.

Additionally there are sometimes so much style issues that it
gets hard to focus on the real issues, e.g. the architecture of a
patch or the overall implementation.

This PR enables a precheck that allows to automatically detect if
our styleguide was followed. The errors can now get cleaned up
before the review, saving time of the reviewer and the submitter.
As most style issues won't appear any more in a review the review
focus will change to a discussion that is more related to the
actual implementation, probably resulting in better reviews and
overall code quality.

PR: #326
PR-URL: https://github.com/apache/couchdb-fauxton/pull/326
Reviewed-By: garren smith <garren.smith@gmail.com>