Mango: creation and listing of indexes
Part 1/2 for Mango:

Creation of Mango indexes and listing them. Disabled pagination
and bulk-deletion for now, see:

https://issues.apache.org/jira/browse/COUCHDB-2651
https://issues.apache.org/jira/browse/COUCHDB-2652

As you can't use the feature right now (querying is part 2/2)
the buttons to reach the features are hidden behind a feature flip.

Use the direct urls to access the features:

http://localhost:8000/#database/$YOUR_DATABASE/_index
http://localhost:8000/#database/$YOUR_DATABASE/_indexlist

Additionally prepares the app for i18n.

Additionally removes the listing of Mango created indexes which
are not editable from the sidebar

COUCHDB-2627