Ace Editor code update
This makes a small change to the Ace Editor component to allow
overriding the default manual JS resizing of the editor. Going
forward we can use flexbox to handle all of that for us, which
will let us drop a lot of the unnecessary JS code to manually
resize the editor based on available space and window resizes,
etc.