Allow customization of view loading styles
By default, the core Fauxton code inserts a spinner for each view
while it's loading. This can be removed via the disableLoader param
but not customized. This PR lets you define an optional
`loaderStyles` property on your view to override the styles. The
use-case here is to provide a light-coloured loader on the dark-
background Doc Editor page.

I investigated using the new React spinner, but ran into problems
targeting views that weren't loaded yet in the page load lifecycle.
This seemed an elegant solution that tapped into what we already
had.