Prevent overlap of text and arrows in styled-select
Super minor thing. If an option in a styled select box
(.styled-select) is quite long, it will overlap the down arrow
overlay which looks kind of bad. Unfortunately browsers are
limited in that you can't put an ellipsis, so this just adds
padding instead to cut off the text, which still looks a lot
better.