Beautify tooltip bug fix
Fixes a small bug where the tooltip for the Beautify button on the
Edit Index page sometimes doesn't disappear after clicking it.
Also moved the tooltip to the right to prevent overlapping, reduced
the button size a little (it was larger than the main buttons on
the page) and removed a wayward semicolon.