General NW robustness test PR
This ticket illustrated there are numerous buggy tests that
fail erratically from time to time:

This is a general ticket to improve the robustness of the NW
tests. I'm going to re-run it again and again and fix whatever
comes up.

- bulkDelete extra check for DOM content
- deleteDatabaseModal.js fix: now ensures the test to confirm the
modal DOM element exists PRIOR to clicking the delete menu item.
- removing test comment
- tweaks to paginateView.js
- ensuring db page is fully loaded before all tests start
- mango test fixes