Prevent duplicate mocha tests running
When running mocha for environments where Fauxton has been
extended, e.g. the Cloudant dashboard, tests unique to that other
environment get run twice because the original tests are included
once from the original folder and once from where they've been
copied - in the Fauxton /app folder. Since we know that the files
would need to be copied over to the /app folder for inclusion
in Fauxton, this PR removes the extra tests from the original
source.