Change default collection to empty collection
This line causes problems when using the component within
other components - not something that currently occurs on Fauxton.
I tried contriving a test but couldn't find an easy way to break
it so decided to forego the test this time. As you can see, all
the current tests work fine.

All it does is change the collection default to an actual Backbone
collection rather than an object.