Zen mode option for code editor
A common complaint is that some of the Ace Editor fields are too
small to really work in, e.g. the Map Function field on the Add/Edit
Index page. This adds an optional Zen Mode option which shows a
fullsize icon next to the fields. That expands the editor to a
full-page editor with limited controls (close + toggle theme from
light to dark).