Refactoring <CodeEditor />
This PR refactors the <CodeEditor /> component to make it
more generic. Ace editor instances such as the mango, index
editor, and Zen mode now use this via wrapper components
(<CodeEditorPanel /> and <ZenModeOverlay />). The goal is
to make this the one any only place where the ace editor
is instantiated. It'll also be needed for the next step:
refactoring the full page doc editor.

Also: Removed extra code to clear notifications in
<Editor />. This was redundant since the "there have been
changed" popup doesn't (and shouldn't, I'd argue) appear
for small editors.