Gruntfile watch task performance improvements
As Fauxton's grown the watch tasks which execute during development
have been taking increasingly longer to run. The old code re-ran
various tasks for every file: JSX generation, copying remote files
over to Fauxton, jshint and stylechecks. When working on a Fauxton
wrapper (like the Dash) often when you save a file and refresh
the page, the various watcher tasks haven't finished executing
so you get errors in the browser. Worse, sometimes the process chokes
and you need to restart the server.

This updated version changes the behavior for JS and JSX files. Now:
- JSX files found outside of the Fauxton folder are compiled
directly into the Fauxton folder.
- JSHint and stylecheck are ran on the changed files only.
- booting up Fauxton (`grunt dev`) behaves just as before and does
a clean slate pulling in the latest content
- JSX module cache disabled. Since it's only ran during the initial
bootup and only takes a fraction of second, disabling it ensures
that on restarting the server everything is completely regenerated
afresh.