Remove create-admin / change password section
CouchDB 2.0 won't support clusterwide configuration. This means
for Fauxton if we use the `/_node` endpoint to change the config
for the password or end admin party by creating a new user we will
talk to a node at some point that requires a password.

As we are probably use a loadbalancer we could try to login now,
but as we don't know if the next request will talk to a node in
admin-party we can't login, at least not in a reliable way.

Once we fixed the cluster-wide-config in a version after 2.0 we
can revert this commit and enable the user section again.

COUCHDB-2343 COUCHDB-2390