csrf: fix tests
when fauxton requests session token & csrf token it handles them
as additional data it sends to each ajax request.

the testsuite tries to open 5984:/_alldocs and friends in several
places right after running the test. as the csrf cookie is sent
via ajax, a plain GET will result in a CSRF error

PR: #494
PR-URL: https://github.com/apache/couchdb-fauxton/pull/494
Reviewed-By: Robert Newson <rnewson@apache.org>
Reviewed-By: Garren Smith <garren.smith@gmail.com>