csrf: fix tests
when fauxton requests session token & csrf token it handles them
as additional data it sends to each ajax request.

the testsuite tries to open 5984:/_alldocs and friends in several
places right after running the test. as the csrf cookie is sent
via ajax, a plain GET will result in a CSRF error