Update Fauxton docs
The current readme is kind of all over the map and a little hard to
follow. This updates it to focus on what new arrivals are most likely
to be interested in: setup and deployment info. Other information
(primarily developer) is now found in a link section at the end
for further reading. See:
https://github.com/benkeen/couchdb-fauxton/blob/update-docs/readme.md