Better handle runtime-loaded files for bundled envs
Occasionally we encounter issues where a file needs to be loaded
at runtime, e.g. ZeroClipboard.swf, Ace editor. To handle this we
have some custom Grunt rules to copy them over to a /js subfolder
where their location is known. This PR removes the custom
code in grunt for Zeroclipboard and make it generic so that
any addon can define a /dependencies subfolder. Anything placed
in that location will be moved to a /js/[addonName] subfolder
for production (bundled) builds.