Catch all errors when failing to load doc
This makes the error handling for loading a doc a little more
generic. This actually caused an issue on the dash, not with
Fauxton - but good to change so that all errors are caught when
failing to load a document and the user's redirected back
to the database all docs page.