Fix for error caused by newlines in View search -> keys field
Curious, very specific bug. When doing a search on a View and entering
a value for the Keys field that contains a newline between two values,
the encoding of the URL gets all messed up and you see a
"missing_named_view" error get thrown within the UI. This ticket simply
strips out newlines in that field.

How to reproduce:
- create a new view
- click on the "Query Options" section and click on Keys to expand
that section. Enter:

["one",
"two"]

Note the newline after the "one" line.
- submit the form.

Result: JS errors and the error notification.