Improve db page error handling
This PR fixes an old bug with the databases page where if one
database query fails, the remaining responses that haven't been
returned don't get populated and the database table has a lot
of incomplete rows.

Now the individual failed database rows don't affect the others and
show a "This database failed to load" message to make it clear to
the user.

The actions.js refactor code was needed due to the following.
FauxtonAPI.when() behaves in an unexpected way: when passed an
array of promises it fires the .always() in two conditions:
1. When all promises are resolved, or,
2. When ONE fails

What was happening was that when a db request failed, it didn't
fire any of the remainder requests, which is why the table would
appear partly empty.

Closes COUCHDB-2243