Add a test to confirm test exists
A pet peeve with running a single mocha test on the command line,
like `grunt file --file=./a/b/c.js` is that (a) it spends several
seconds compiling, copying etc. then (b) just ends. It doesn't
notify you that the file doesn't exist.

This little PR adds a test that runs up front: if the file you
entered doesn't exist, it lets you know about it and stops the
process immediately.