Grunt mocha test file existence check
This is a small tweak to a recent commit that does an up-front
check to confirm the mocha test exists. It had failed to check
the condition where the user was trying to run a .js mocha test
for a ,jsx file that hadn't yet been compiled. This expands on the
existence check for a .jsx version of the file.