Fix for logout link showing up twice in footer
An old bug is the Logout link in the footer showing up twice. It's
a little difficult to reproduce consistently, but I've found this
works the best (Mac, chrome):
1. open dev tools
2. Login to Fauxton, notice the single "logout" link in footer.
3. click on the browser URL bar and click <enter>
4. Repeat the above until you see two "logout" links in the footer.

What's happening is the change event on Auth.session is firing
twice, each of which adds a new footer link. Seems like with
some page loads it only fires once, sometimes twice.

This ensures the link is removed prior to adding it.