Increase height of clickable space for close notification
There's a little "x" that appears in the notification center next
to each message that allows users to remove that single message.
On Firefox, the height of the clickable space wasn't quite large
enough so if you clicked the bottom of the x it wouldn't close.
This fixes it so that it covers the full "x".