Configurable 'Generated: [date]' comments
A while back we added in a comment to the foot of the generated
index.html file containing a "Generated: [data]" text. This has
proven to be extremely useful.

This PR just expands on it a little to make the label itself
configurable. So for environments using different bundles of
the Fauxton code, it can output the build target name, plus
the generation date, e.g.

<!-- Env X. Generated: [date] -->