Sidebar component updates
This makes a couple of changes to the DesignDoc and IndexSection
components used in the sidebar.

1. `<IndexSection />` knew too much! It had some hardcoded
references to the structure (prop names) of the design docs and
how it was being called. Now it’s more generic and passed
everything it needs to know via props, so it’s a little bit
dumber + will work for any new type of design doc contents.

2. The design doc item in the sidebar will now only list
Search Indexes or other subsections if they have content. This
won’t impact Fauxton as is, but for script that wrap this
code, it’ll cut down on visual clutter in the UI. [This is
mostly for UX.]