Minor CSS tweaks
After speaking with Justin, I made a small tweak to the
header CSS which ensures the spacing between the API and
OPTIONS header buttons and their icons was consistent
regardless of viewport size. That lead to a few other small
CSS tweaks to ensure it works on all screen sizes.