remove warnings on console
the warnings are displayed per default even for a valid release.

the warnings confuse people as they things something is broken,
but it is not.

this leads to a mindset were warnings are generally ignored as
the warning don't say anything about the status of the
application. this patch removes the warnings as debug logging can
get inserted again when there is something to debug.

PR: #625
PR-URL: https://github.com/apache/couchdb-fauxton/pull/625
Reviewed-By: Benjamin Keen <ben.keen@gmail.com>