fontawesome publishing changes
This PR handles a bit of regression that always occurs when
we regenerate the font icon files.

This'll take a bit of explanation...

On Fauxton, we use a custom Ruby template to generate our
icons.less file, see:
https://github.com/apache/couchdb-fauxton/blob/master/assets/fonts/templates/icons.less

The problem with this is that the generated CSS is of the form:

.icon1-classname:before { content: "\f123"; }
.icon2-classname:before { content: "\f234"; }
...

That's perfectly fine CSS, but because the class isn't defined
without the :before pseudo class after it, we can't reference
and use that class in other classes in our LESS files. This
is problematic because any time we either have to re-use the
class then override everything we don't want (which I've found
wasn't always possible) or we hardcode the unicode char, like:
.mynewclass { content: "\u123"; }

And THAT is a problem because any time we update the list of
fonts and regenerate the font files, the unicode char assigned
to that class may change. As such, we have to re-examine the
whole codebase to update any icons that may have changed.

Sooo..... this adds a new glyph_less var that generates the output
in a Less-friendly format:
.icon1-classname { &:before { ... }}

Ultimately we should replace the whole lib or update it all,
but I have an immediate need to add new icons and don't want to
manually fix things yet again.