Update Views
This is a big update to the Views pages.
- now uses standard 2 panel layout for adding/editing/viewing
Views (includes the sidebar)
- New menu added to all Views, containing [Edit, Clone, Delete]
options. Clone and Delete are their own modals. This modal is
extensible, so can be used for different types of indexes added
via extensions.
- sidebar now properly selects views and highlights them to
show current location + view.
- Renaming a view no longer creates a second view, just renames
it. The clone feature lets you copy/clone a View. It also properly
removes the old view from the old location, be it a separate
design doc or the current one. If it removes it from another
design doc, it deletes the old design doc if it's empty.
- Auto-deleting logic of design docs has been rewritten. Before,
we were overloading the hasViews() method on the backbone object
to determine if the design doc was now empty (this didn't
actually work, btw). Now any code that extends Fauxton's index
types, registers the design doc properly name for the index
groups. That is then used in the deletion code to properly
identify an empty design doc.

N.B. the new menu uses React Bootstrap so it's very similar
but not identical to our current menu dropdowns. The latter
will be replaced in subsquent PRs.