Make DesignDocSelector into a dumb component
The dumbifies the DesignDocSelector to remove all ties a to
store and passes everything via props so we can get more use out
of it. Specifically, I need this for my next PR which includes
a new, generic Clone Index modal that'll use this component
in different contexts (i.e. all index types).

- Includes validation in the component to make it more
self-contained and allows for better UX by focusing on error
fields.
- Save action significantly simplied to remove custom update
logic and always redirect to appropriate View page.