Revert "JS validation added for Add Database"
This reverts commit b8bbb87c6919469521ff277694fb75e898157a7c.

The commit was added in https://github.com/apache/couchdb-fauxton/pull/468
but the validation has to happen in the backend. The current
validation does not cover databases which are created via curl or
other apps. It also prevents users from creating an accidentally
deleted `_users` database.

The fix for the root cause has to happen in the backend.

We should avoid trying to add validations in multiple, different
places.

COUCHDB-2748