Editor: don't throw if document has no id
Right now the editor throws if you try to create a doc that does
nto have an id, instead of doing a POST to the database, which
will cause CouchDB to create an id for us.

Sometimes you want to be able to not use the pre-generated id from
Fauxton. Use cases are copy pasting a JSON object into the editor
or if you want to let the server take care of creating the id.

This patch lets fixes an invalid POST to `$DB/undefined` in cases
where the id wasn't specified.

Thanks to Will Holley for reporting!

PR: #678
PR-URL: https://github.com/apache/couchdb-fauxton/pull/678
Reviewed-By: garren smith <garren.smith@gmail.com>