Revert to version 1.1.2
hello,

I was updating the npm fauxton release today, and published the
tag 1.1.3.

Usually my next step would be to update the rebar.config file in
couchdb, but when I tested the new release (reinstalling it
globally), then fauxton, then localhost:8000, it wasn’t working,
so i reset 1.1.2 as the latest release.
(so that if anyone npm installs it today, it will work.)

I'm changing the version number back to match the version on
rebar.config and npm.