Style fix for doc ID block
When you typed in something in the document ID typeahead field
it would look a little odd if there were no results (the grey
block would be fixed at 291px). This changes that to a max height
so it only ever appears as tall as it needs to be (and looks nice
when you're typing and honing in on a doc).

It also adjusts the width to fill as much room as is available
when used in the header.