Move selenium to docker
This runs selenium and CouchDB in a docker instance for the nightwatch
tests. It also includes a selenium debug container that allows you to
view the tests via VNC. This change is to make the running of the
selenium tests more consistent for all developers and the same as on
Travis.