remove update sequence from db list
right now the update sequence display is broken. they also make
no sense in a clustered setup.

this patch removes them test-wise. we can readd them if there is
a lot demand for seeing them, but as we got no bug reports in
almost 2 years it seems it is a dead feature.

PR: #776
PR-URL: https://github.com/apache/couchdb-fauxton/pull/776
Reviewed-By: garren smith <garren.smith@gmail.com>