(#868) - Update to webpack 2
This upgrades to webpack 2. Webpack 2 makes it a little easier to do certain things. One of them is to split the bundle.js into three separate files:

manifest.js -> webpack loader file
vendor.js -> all required libraries
bundle.js -> Fauxton code
Each file also has a a hash applied to it based on the file contents. This should improve caching and load times for Fauxton as the manifest file and vendor file will not change as often as the bundle file.

Webpack is now also generating the html file, this was previously done in grunt.