Update for changes to _cluster_setup endpoint
These changes add support for the new enable_single_node action,
which tolerates binding a node to 127.0.0.1.

They also add support for the new mandatory node_count parameter
when setting up a cluster. This parameter in the API call actually
sets [cluster] n value, so it is set to min(3, <value provided>)
based on recommendations from @rnewson and personal experience.

Closes #874