Show a warning in Fauxton when no Mango index is available
Mango allows queries to run without an index by performing a full
database scan. Whilst this is useful in some scenarios, Couch
returns a warning in the response that an index should be created
if performance is important. This warning is now propagated
to Fauxton using the existing alert infrastructure.