Display execution statistics for Query
When running a Mango Query, display the execution stats and any
warnings returned. The query execution time and presence of
a warning is indicated in the query editor after a query
has been run. An overlay popup displays the detailed
execution statistics and warning text.

Not all versions of CouchDB support execution statistics for Mango,
so detect that the feature is supported before requesting them
at query time.

In the case that execution statistics are not available, we still
display any warnings that are returned.